[printer] Refactor printer into its own package
