+++ samlang-cli/package.json
@@ -14,12 +14,12 @@
     "samlang": "./bin/index.js"
   },
   "devDependencies": {
-    "@dev-sam/samlang-core": "workspace:0.0.1",
     "@types/jest": "^26.0.10",
     "@types/node": "^14.6.0",
     "@vercel/ncc": "^0.24.0",
     "samlang-core-ast": "workspace:0.0.1",
     "samlang-core-compiler": "workspace:0.0.1",
+    "samlang-core-printer": "workspace:0.0.1",
     "samlang-core-services": "workspace:0.0.1",
     "typescript": "4.0.2",
     "vscode-languageserver": "^6.1.1"
