+++ samlang-core-printer/package.json
@@ -1,24 +1,21 @@
 {
-  "name": "@dev-sam/samlang-core",
+  "name": "samlang-core-printer",
   "version": "0.0.1",
   "license": "AGPLv3",
   "scripts": {
     "compile": "tsc --incremental --noEmit"
   },
-  "devDependencies": {
-    "@types/jest": "^26.0.10",
-    "typescript": "4.0.2"
-  },
   "dependencies": {
     "samlang-core-ast": "workspace:0.0.1",
+    "samlang-core-utils": "workspace:0.0.1"
+  },
+  "devDependencies": {
+    "@types/jest": "^26.0.10",
     "samlang-core-checker": "workspace:0.0.1",
     "samlang-core-compiler": "workspace:0.0.1",
     "samlang-core-errors": "workspace:0.0.1",
-    "samlang-core-interpreter": "workspace:0.0.1",
-    "samlang-core-optimization": "workspace:0.0.1",
     "samlang-core-parser": "workspace:0.0.1",
-    "samlang-core-printer": "workspace:0.0.1",
     "samlang-core-services": "workspace:0.0.1",
-    "samlang-core-utils": "workspace:0.0.1"
+    "typescript": "4.0.2"
   }
 }
