+++ samlang-core-printer/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "extends": "../configurations/tsconfig.base.json"
+}
