+++ samlang-demo/package.json
@@ -12,12 +12,12 @@
     "bundle": "node build.js"
   },
   "devDependencies": {
-    "@dev-sam/samlang-core": "workspace:0.0.1",
     "@types/jest": "^26.0.10",
     "@vercel/ncc": "^0.24.0",
     "samlang-core-ast": "workspace:0.0.1",
     "samlang-core-compiler": "workspace:0.0.1",
     "samlang-core-interpreter": "workspace:0.0.1",
+    "samlang-core-printer": "workspace:0.0.1",
     "samlang-core-services": "workspace:0.0.1",
     "typescript": "4.0.2"
   }
