+++ samlang-demo/src/index.ts
@@ -1,8 +1,8 @@
-import { prettyPrintSamlangModule, prettyPrintHighIRModuleAsJS } from '@dev-sam/samlang-core';
 import { assemblyProgramToString } from 'samlang-core-ast/asm-program';
 import { ModuleReference } from 'samlang-core-ast/common-nodes';
 import { compileSamlangSourcesToHighIRSources } from 'samlang-core-compiler';
 import interpretSamlangModule from 'samlang-core-interpreter/source-level-interpreter';
+import { prettyPrintSamlangModule, prettyPrintHighIRModuleAsJS } from 'samlang-core-printer';
 import { checkSources, lowerSourcesToAssemblyPrograms } from 'samlang-core-services';
 
 type SamlangDemoResult = {
