+++ workspaces.json
@@ -6,15 +6,14 @@
       "dependencyChain": [
         "samlang-core-utils",
         "samlang-core-ast",
-        "samlang-core-errors",
-        "samlang-core-parser",
-        "samlang-core-checker",
         "samlang-core-analysis",
         "samlang-core-optimization",
         "samlang-core-compiler",
-        "samlang-core-interpreter",
+        "samlang-core-errors",
+        "samlang-core-parser",
+        "samlang-core-checker",
         "samlang-core-services",
-        "@dev-sam/samlang-core",
+        "samlang-core-printer",
         "@dev-sam/samlang-cli"
       ]
     },
@@ -31,6 +30,7 @@
         "samlang-core-compiler",
         "samlang-core-interpreter",
         "samlang-core-services",
+        "samlang-core-printer",
         "@dev-sam/samlang-core"
       ]
     },
@@ -39,15 +39,15 @@
       "dependencyChain": [
         "samlang-core-utils",
         "samlang-core-ast",
-        "samlang-core-errors",
-        "samlang-core-parser",
-        "samlang-core-checker",
         "samlang-core-analysis",
         "samlang-core-optimization",
         "samlang-core-compiler",
+        "samlang-core-errors",
+        "samlang-core-parser",
         "samlang-core-interpreter",
+        "samlang-core-checker",
         "samlang-core-services",
-        "@dev-sam/samlang-core",
+        "samlang-core-printer",
         "@dev-sam/samlang-demo"
       ]
     },
@@ -122,6 +122,21 @@
         "samlang-core-parser"
       ]
     },
+    "samlang-core-printer": {
+      "workspaceLocation": "samlang-core-printer",
+      "dependencyChain": [
+        "samlang-core-utils",
+        "samlang-core-ast",
+        "samlang-core-errors",
+        "samlang-core-parser",
+        "samlang-core-checker",
+        "samlang-core-analysis",
+        "samlang-core-optimization",
+        "samlang-core-compiler",
+        "samlang-core-services",
+        "samlang-core-printer"
+      ]
+    },
     "samlang-core-services": {
       "workspaceLocation": "samlang-core-services",
       "dependencyChain": [
@@ -146,16 +161,17 @@
   "topologicallyOrdered": [
     "samlang-core-utils",
     "samlang-core-ast",
-    "samlang-core-errors",
-    "samlang-core-parser",
-    "samlang-core-checker",
     "samlang-core-analysis",
     "samlang-core-optimization",
     "samlang-core-compiler",
-    "samlang-core-interpreter",
+    "samlang-core-errors",
+    "samlang-core-parser",
+    "samlang-core-checker",
     "samlang-core-services",
-    "@dev-sam/samlang-core",
+    "samlang-core-printer",
     "@dev-sam/samlang-cli",
+    "samlang-core-interpreter",
+    "@dev-sam/samlang-core",
     "@dev-sam/samlang-demo"
   ]
 }
